refactor(#364): 프로필 셀 코드 리팩토링
